Small modifications to allow building on NUMA machine
The first change here is modifying the makefile to use g++-5 instead
of whatever the default version of g++ is. g++-5 is necessary for a
number of modern c++ features, but changing the global default install
on Jose's machine could potentially cause problems for him, so
specifying the version explicitly seems like a good solution.

The second is deleting the gtest.h header from version control. It
doesn't really make sense to have the header stored locally when the
rest of the library is installed globally, and this can cause problems
when one is updated without updating the other.